main file started to implement

+++ init_config.c
@@ -15,8 +15,8 @@ init_config3: alternate the entries between 1 and -1
 void init_config1(char *configuration, parameters params){
 	long int L= params.L;
 	char *configuration = malloc(params.L * sizeof(char));
-    
-    for (int i=0; i<L; i++) {
+
+    for (long int i=0; i<L; i++) {
         if((double) rand()/RAND_MAX<0.5){
             configuration[i]=1;
         }
@@ -31,8 +31,8 @@ void init_config2(char *configuration, parameters params){
 	gsl_rng * r = gsl_rng_alloc (gsl_rng_taus);
 	long int r_max = gsl_rng_max(r)
 	char *configuration = malloc(params.L * sizeof(char));
-    
-    for (int i=0; i<L; i++) {
+
+    for (long int i=0; i<L; i++) {
         if(unsigned long int gsl_rng_get(r)/r_max<0.5){
             configuration[i]=1;
         }
@@ -45,11 +45,11 @@ void init_config2(char *configuration, parameters params){
 void init_config3(char *configuration, parameters params){
 	long int L= params.L;
 	char *configuration = malloc(params.L * sizeof(char));
-    
-    for (int i=0; i<L; i=i+2) {
+
+    for (long int i=0; i<L; i=i+2) {
         init_config[i]=1;
     }
-    for (int i=1; i<L; i=i+2) {
+    for (long int i=1; i<L; i=i+2) {
         configuration[i]=-1;
     }
 }
